Extract secur32 interop for netfx usage only
