+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft.Data.SqlClient.csproj
@@ -151,6 +151,15 @@
     <Compile Include="$(CommonSourceRoot)Interop\Windows\SystemErrors.cs">
       <Link>Interop\SystemErrors.cs</Link>
     </Compile>
+    <Compile Include="$(CommonSourceRoot)\Interop\Windows\Secur32\CredHandle.netfx.cs">
+      <Link>Interop\Windows\Secur32\CredHandle.netfx.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonSourceRoot)\Interop\Windows\Secur32\ContextAttribute.netfx.cs">
+      <Link>Interop\Windows\Secur32\ContextAttribute.netfx.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonSourceRoot)\Interop\Windows\Secur32\Secur32.netfx.cs">
+      <Link>Interop\Windows\Secur32\Secur32.netfx.cs</Link>
+    </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Sni\AuthProviderInfo.cs">
       <Link>Interop\Windows\Sni\AuthProviderInfo.cs</Link>
     </Compile>
