+++ src/Microsoft.Data.SqlClient/src/Interop/Windows/Secur32/CredHandle.netfx.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#if NETFRAMEWORK
+
+using System;
+using System.Runtime.InteropServices;
+
+namespace Interop.Windows.Secur32
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal struct CredHandle
+    {
+        internal IntPtr dwLower;
+        internal IntPtr dwUpper;
+    };
+}
+
+#endif
